+++ Exception/InvalidTransformationException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Tdn\PhpTypes\Exception;
+
+class InvalidTransformationException extends \RuntimeException
+{
+}