+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Unreleased
+## 2.0.0 - 2015-09-08
 
 * Upgraded libs
 * Removed indent functions. Fringe case.